+++ java-security-it/src/test/java/com/sap/cloud/security/test/performance/SpringSecurityPerformanceIT.java
@@ -80,7 +80,7 @@ class SpringSecurityPerformanceIT {
 
 	private OAuth2ServiceConfigurationBuilder createIasConfigurationBuilder() {
 		return OAuth2ServiceConfigurationBuilder.forService(IAS)
-				.withDomains(SecurityTest.DEFAULT_DOMAIN)
+				.withDomains(securityIasTest.getWireMockServer().baseUrl())
 				.withClientId(SecurityTest.DEFAULT_CLIENT_ID);
 	}
 }